TestFloat Initial commit
This is testfloat from http://www.jhauser.us/arithmetic/TestFloat.html

This is 2c since I need it for testing some things in gdb.  I have moved
things around and changed the make file a bit to make it easier for me
to maintain and a bit more similar to other works I am fimilar with.